+++ tests_phpunit/src/Runtime/RunContextStoreFileTest.php
@@ -68,4 +68,17 @@ class RunContextStoreFileTest extends TestCase {
     $this->assertEquals($data, $store->get('key1'));
   }
 
+  public function testRemove() {
+    $store = new RunContextStoreFile($this->filePath);
+    $store->set('key1', 'value1');
+    $this->assertTrue($store->has('key1'));
+    $store->remove('key1');
+    $this->assertFalse($store->has('key1'));
+    $this->assertNull($store->get('key1'));
+
+    // Test persistence
+    $store_two = new RunContextStoreFile($this->filePath);
+    $this->assertFalse($store_two->has('key1'));
+  }
+
 }